# docs/swagger.yaml
basePath: /api/v1
definitions:
  request.ChangePasswordRequest:
    properties:
      new_password:
        minLength: 8
        type: string
    required:
    - new_password
    type: object
  response.ErrorSwaggerResponse:
    properties:
      errors:
        type: string
      message:
        example: error
        type: string
      success:
        example: false
        type: boolean
    type: object
  response.SuccessSingleUserResponse:
    properties:
      data:
        $ref: '#/definitions/user.User'
      message:
        example: success
        type: string
      success:
        example: true
        type: boolean
    type: object
  response.SuccessUserResponse:
    properties:
      data:
        items:
          $ref: '#/definitions/user.User'
        type: array
      message:
        example: success
        type: string
      success:
        example: true
        type: boolean
    type: object
  user.LoginRequest:
    properties:
      email:
        type: string
      password:
        type: string
    required:
    - email
    - password
    type: object
  user.User:
    properties:
      email:
        type: string
      id:
        type: string
      is_active:
        type: boolean
      name:
        type: string
      roles:
        items:
          type: string
        type: array
    type: object
host: localhost:8080
info:
  contact:
    email: mohamadafandi71@gmail.com
    name: Afandi
  description: Clean Architecture Go Gin API
  termsOfService: http://swagger.io/terms/
  title: Go Gin API
  version: "1.0"
paths:
  /login:
    post:
      consumes:
      - application/json
      parameters:
      - description: Login payload
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/user.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.SuccessSingleUserResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ErrorSwaggerResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.ErrorSwaggerResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.ErrorSwaggerResponse'
      summary: Login user
      tags:
      - Users
  /users:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.SuccessUserResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.ErrorSwaggerResponse'
      summary: Get all users
      tags:
      - Users
    post:
      consumes:
      - application/json
      parameters:
      - description: User payload
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/user.User'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/response.SuccessSingleUserResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ErrorSwaggerResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.ErrorSwaggerResponse'
      summary: Create user
      tags:
      - Users
  /users/{id}:
    delete:
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.SuccessSingleUserResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ErrorSwaggerResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.ErrorSwaggerResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.ErrorSwaggerResponse'
      summary: Delete user
      tags:
      - Users
    get:
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.SuccessSingleUserResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.ErrorSwaggerResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.ErrorSwaggerResponse'
      summary: Get user by ID
      tags:
      - Users
    put:
      consumes:
      - application/json
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: integer
      - description: User payload
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/user.User'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.SuccessSingleUserResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ErrorSwaggerResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.ErrorSwaggerResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.ErrorSwaggerResponse'
      summary: Update user
      tags:
      - Users
  /users/{id}/change-password:
    put:
      consumes:
      - application/json
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: integer
      - description: Change password payload
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/request.ChangePasswordRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.SuccessSingleUserResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ErrorSwaggerResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.ErrorSwaggerResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.ErrorSwaggerResponse'
      summary: Change user password
      tags:
      - Users
swagger: "2.0"